requirements.txt
picamera
numpy
scipy
python-dotenv
python-telegram-bot
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pillow
